add implementation of dao and interfaces

+++ ws-core/src/main/java/com/epam/lab/ticket/manager/dao/impl/hibernate/HibernateTicketDao.java
@@ -1,8 +1,11 @@
 package com.epam.lab.ticket.manager.dao.impl.hibernate;
 
 import com.epam.lab.ticket.manager.dao.TicketDao;
+import com.epam.lab.ticket.manager.entity.StateTicket;
 import com.epam.lab.ticket.manager.entity.Ticket;
 import com.epam.lab.ticket.manager.exception.DaoException;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
@@ -13,18 +16,28 @@ import org.springframework.stereotype.Repository;
 @Repository
 public class HibernateTicketDao implements TicketDao {
 
+    private static final Logger logger = LogManager.getLogger(HibernateTicketDao.class);
+
     @Autowired
     private SessionFactory factory;
 
     @Override
-    public boolean payForTicket(Ticket ticket) throws DaoException {
+    public boolean updateStatusOfTicket(Long idTicket) throws DaoException {
+        boolean result = false;
         Session session = factory.getCurrentSession();
         try {
-            session.saveOrUpdate(ticket);
+            Ticket ticket = (Ticket) session.get(Ticket.class, idTicket);
+            if (ticket != null) {
+                ticket.setState(StateTicket.BOUGHT);
+                session.update(ticket);
+                result = true;
+            } else {
+                logger.warn("Ticket with id= " + idTicket.toString() + " not found");
+            }
         } catch (HibernateException e) {
             throw new DaoException(e);
         }
-        return true;
+        return result;
     }
 
     @Override
@@ -53,15 +66,21 @@ public class HibernateTicketDao implements TicketDao {
 
     @Override
     public boolean delete(Long key) throws DaoException {
+        boolean result;
         try {
             Session session = factory.getCurrentSession();
-            Ticket ticket = new Ticket();
-            ticket.setNumberTicket(key);
-            session.delete(ticket);
+            Ticket ticket = (Ticket) session.get(Ticket.class, key);
+            if (ticket != null) {
+                session.delete(ticket);
+                result = true;
+            } else {
+                logger.warn("Ticket with id " + key.toString() + "not found");
+                result = false;
+            }
         } catch (HibernateException e) {
             throw new DaoException("Error delete ticket with id = " + key, e);
         }
-        return true;
+        return result;
     }
 
     @Autowired
